renamed app to database and moved to server folder, added routes for item api call, fixed config require on item and wishlist models
