+++ server/routing.js
@@ -1,8 +1,7 @@
 var bodyParser = require('body-parser');
 var path = require('path');
 var wishlistController = require('./wishlist/wishlistController');
-
-// TODO: Require wishlistController here.
+var itemController = require('./item/itemController');
 
 module.exports = function (app, express) {
   app.use(bodyParser.urlencoded({ extended: true }));
@@ -10,5 +9,10 @@ module.exports = function (app, express) {
   app.use(express.static(path.join(__dirname, '/../client')));
   app.use(express.static(path.join(__dirname, '/../node_modules')));
 
+  //requests for wishlists
   app.get('/api/wishlist', wishlistController.getAll);
+  app.post('/api/wishlist', wishlistController.addOne);
+  //requests for items
+  app.get('/api/item', itemController.getAll);
+  app.post('/api/item', itemController.addOne);
 };
